feat: add the ability to skip storage of keystores/passwords for keymanager keys
This PR introduces a new flag: --Xkey-manager-skip-keystore-storage

There is currently no way to prevent web3signer to write keystores and
corresponding passwords on-disk using the keymanager API. This PRs
adds a flag to prevent this from happening, which reduces an attack
vector (i.e: access to the filesystem of web3signer). This also makes
it possible to harden the setup of Web3signer in some environment via
read-only devices.

This is *not* safe for operators who do not backup keystore.